fixing information route
